OPIK-1386 refactor projects logic out of FeedbackScoreService
