+++ apps/opik-backend/src/main/java/com/comet/opik/domain/FeedbackScoreService.java
@@ -4,7 +4,6 @@ import com.comet.opik.api.FeedbackScore;
 import com.comet.opik.api.FeedbackScoreBatchItem;
 import com.comet.opik.api.FeedbackScoreNames;
 import com.comet.opik.api.Project;
-import com.comet.opik.infrastructure.auth.RequestContext;
 import com.comet.opik.utils.BinaryOperatorUtils;
 import com.comet.opik.utils.WorkspaceUtils;
 import com.google.inject.ImplementedBy;
@@ -13,13 +12,9 @@ import jakarta.inject.Inject;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.jdbi.v3.core.statement.UnableToExecuteStatementException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
-import ru.vyarus.guicey.jdbi3.tx.TransactionTemplate;
 
-import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -28,9 +23,6 @@ import java.util.UUID;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static com.comet.opik.domain.FeedbackScoreDAO.EntityType;
-import static com.comet.opik.infrastructure.db.TransactionTemplateAsync.READ_ONLY;
-import static com.comet.opik.infrastructure.db.TransactionTemplateAsync.WRITE;
 import static com.comet.opik.utils.ErrorUtils.failWithNotFound;
 import static java.util.stream.Collectors.groupingBy;
 
@@ -61,10 +53,9 @@ public interface FeedbackScoreService {
 class FeedbackScoreServiceImpl implements FeedbackScoreService {
 
     private final @NonNull FeedbackScoreDAO dao;
-    private final @NonNull TransactionTemplate syncTemplate;
     private final @NonNull SpanDAO spanDAO;
     private final @NonNull TraceDAO traceDAO;
-    private final @NonNull IdGenerator idGenerator;
+    private final @NonNull ProjectService projectService;
 
     record ProjectDto(Project project, List<FeedbackScoreBatchItem> scores) {
     }
@@ -114,30 +105,10 @@ class FeedbackScoreServiceImpl implements FeedbackScoreService {
                 })
                 .collect(groupingBy(FeedbackScoreBatchItem::projectName));
 
-        return handleProjectRetrieval(scoresPerProject)
+        return projectService.retrieveByNamesOrCreate(scoresPerProject.keySet())
                 .map(this::groupByName)
                 .map(projectMap -> mergeProjectsAndScores(projectMap, scoresPerProject))
                 .flatMap(projects -> processScoreBatch(entityType, projects, scores.size())) // score all scores
-                .then();
-    }
-
-    private Mono<List<Project>> handleProjectRetrieval(Map<String, List<FeedbackScoreBatchItem>> scoresPerProject) {
-        return Mono.deferContextual(ctx -> {
-            String workspaceId = ctx.get(RequestContext.WORKSPACE_ID);
-            String userName = ctx.get(RequestContext.USER_NAME);
-
-            return checkIfNeededToCreateProjectsWithContext(workspaceId, userName, scoresPerProject) // create projects if needed
-                    .then(Mono.fromCallable(() -> getAllProjectsByName(workspaceId, scoresPerProject))
-                            .subscribeOn(Schedulers.boundedElastic())); // get all project itemIds
-        });
-    }
-
-    private Mono<Void> checkIfNeededToCreateProjectsWithContext(String workspaceId,
-            String userName,
-            Map<String, List<FeedbackScoreBatchItem>> scoresPerProject) {
-
-        return Mono.fromRunnable(() -> checkIfNeededToCreateProjects(scoresPerProject, userName, workspaceId))
-                .publishOn(Schedulers.boundedElastic())
                 .then();
     }
 
@@ -171,54 +142,6 @@ class FeedbackScoreServiceImpl implements FeedbackScoreService {
                 Function.identity(),
                 BinaryOperatorUtils.last(),
                 () -> new TreeMap<>(String.CASE_INSENSITIVE_ORDER)));
-    }
-
-    private List<Project> getAllProjectsByName(String workspaceId,
-            Map<String, List<FeedbackScoreBatchItem>> scoresPerProject) {
-        return syncTemplate.inTransaction(READ_ONLY, handle -> {
-
-            var projectDAO = handle.attach(ProjectDAO.class);
-
-            return projectDAO.findByNames(workspaceId, scoresPerProject.keySet());
-        });
-    }
-
-    private void checkIfNeededToCreateProjects(Map<String, List<FeedbackScoreBatchItem>> scoresPerProject,
-            String userName, String workspaceId) {
-
-        Map<String, Project> projectsPerLowerCaseName = groupByName(
-                getAllProjectsByName(workspaceId, scoresPerProject));
-
-        syncTemplate.inTransaction(WRITE, handle -> {
-
-            var projectDAO = handle.attach(ProjectDAO.class);
-
-            scoresPerProject
-                    .keySet()
-                    .stream()
-                    .filter(projectName -> !projectsPerLowerCaseName.containsKey(projectName))
-                    .forEach(projectName -> {
-                        UUID projectId = idGenerator.generateId();
-                        var newProject = Project.builder()
-                                .name(projectName)
-                                .id(projectId)
-                                .createdBy(userName)
-                                .lastUpdatedBy(userName)
-                                .build();
-
-                        try {
-                            projectDAO.save(workspaceId, newProject);
-                        } catch (UnableToExecuteStatementException e) {
-                            if (e.getCause() instanceof SQLIntegrityConstraintViolationException) {
-                                log.warn("Project {} already exists", projectName);
-                            } else {
-                                throw e;
-                            }
-                        }
-                    });
-
-            return null;
-        });
     }
 
     @Override
