+++ apps/opik-backend/src/main/java/com/comet/opik/domain/ProjectService.java
@@ -17,6 +17,7 @@ import com.comet.opik.domain.sorting.SortingQueryBuilder;
 import com.comet.opik.domain.stats.StatsMapper;
 import com.comet.opik.infrastructure.auth.RequestContext;
 import com.comet.opik.infrastructure.db.TransactionTemplateAsync;
+import com.comet.opik.utils.BinaryOperatorUtils;
 import com.comet.opik.utils.PaginationUtils;
 import com.google.inject.ImplementedBy;
 import jakarta.inject.Inject;
@@ -29,6 +30,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.jdbi.v3.core.statement.UnableToExecuteStatementException;
+import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 import ru.vyarus.guicey.jdbi3.tx.TransactionTemplate;
 
 import java.sql.SQLIntegrityConstraintViolationException;
@@ -40,6 +43,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.TreeMap;
 import java.util.UUID;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -82,6 +86,8 @@ public interface ProjectService {
     Project getOrCreate(String workspaceId, String projectName, String userName);
 
     Project retrieveByName(String projectName);
+
+    Mono<List<Project>> retrieveByNamesOrCreate(Set<String> projectNames);
 
     void recordLastUpdatedTrace(String workspaceId, Collection<ProjectIdLastUpdated> lastUpdatedTraces);
 
@@ -499,9 +505,83 @@ class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
+    public Mono<List<Project>> retrieveByNamesOrCreate(Set<String> projectNames) {
+        return Mono.deferContextual(ctx -> {
+            String workspaceId = ctx.get(RequestContext.WORKSPACE_ID);
+            String userName = ctx.get(RequestContext.USER_NAME);
+
+            return checkIfNeededToCreateProjectsWithContext(workspaceId, userName, projectNames) // create projects if needed
+                    .then(Mono.fromCallable(() -> getAllProjectsByName(workspaceId, projectNames))
+                            .subscribeOn(Schedulers.boundedElastic())); // get all project itemIds
+        });
+    }
+
+    @Override
     public void recordLastUpdatedTrace(String workspaceId, Collection<ProjectIdLastUpdated> lastUpdatedTraces) {
         template.inTransaction(WRITE,
                 handle -> handle.attach(ProjectDAO.class).recordLastUpdatedTrace(workspaceId, lastUpdatedTraces));
     }
 
+    private Mono<Void> checkIfNeededToCreateProjectsWithContext(String workspaceId,
+            String userName, Set<String> projectNames) {
+
+        return Mono.fromRunnable(() -> checkIfNeededToCreateProjects(projectNames, userName, workspaceId))
+                .publishOn(Schedulers.boundedElastic())
+                .then();
+    }
+
+    private List<Project> getAllProjectsByName(String workspaceId,
+            Set<String> projectNames) {
+        return template.inTransaction(READ_ONLY, handle -> {
+
+            var projectDAO = handle.attach(ProjectDAO.class);
+
+            return projectDAO.findByNames(workspaceId, projectNames);
+        });
+    }
+
+    private void checkIfNeededToCreateProjects(Set<String> projectNames,
+            String userName, String workspaceId) {
+
+        Map<String, Project> projectsPerLowerCaseName = groupByName(
+                getAllProjectsByName(workspaceId, projectNames));
+
+        template.inTransaction(WRITE, handle -> {
+
+            var projectDAO = handle.attach(ProjectDAO.class);
+
+            projectNames
+                    .stream()
+                    .filter(projectName -> !projectsPerLowerCaseName.containsKey(projectName))
+                    .forEach(projectName -> {
+                        UUID projectId = idGenerator.generateId();
+                        var newProject = Project.builder()
+                                .name(projectName)
+                                .id(projectId)
+                                .createdBy(userName)
+                                .lastUpdatedBy(userName)
+                                .build();
+
+                        try {
+                            projectDAO.save(workspaceId, newProject);
+                        } catch (UnableToExecuteStatementException e) {
+                            if (e.getCause() instanceof SQLIntegrityConstraintViolationException) {
+                                log.warn("Project {} already exists", projectName);
+                            } else {
+                                throw e;
+                            }
+                        }
+                    });
+
+            return null;
+        });
+    }
+
+    private Map<String, Project> groupByName(List<Project> projects) {
+        return projects.stream().collect(Collectors.toMap(
+                Project::name,
+                Function.identity(),
+                BinaryOperatorUtils.last(),
+                () -> new TreeMap<>(String.CASE_INSENSITIVE_ORDER)));
+    }
 }
